Improve EventStoreSubscriber stability.
In some weird situations, null reference exceptions were being thrown when restarting a subscriber, and preventing restarts; this has been fixed. The EventStoreSubscriber was also not trying to restart in situations where it should have, which has been fixed, and subscriptions were occasionally being duplicated; to prevent this, every time the Subscription is started and restarted, inside of the lock which contains the start logic, it is now ensured that no other subscription exists.